Rewrote loops without type casts
The syntax `for(A;B;C){E;}` is equivalent to `A; while(B){E; C;}`. Hence `for(i=D; i-->0;){E;}` sets `i` to `D`, compares it strictly to `0`, then decrements `i` to `D-1` before executing the loop code `E` for the first time. The last execution of the loop code `E` is with `i=0`.