Fix kdv_discretization_2SPLIT2A
This scheme uses a slightly different basis than AKNS in order to reduce the order of the polynomials for each sample by 1. This is (still) not implemented in nsev.c, from which the new kdvv.c was copied. In effect this reduces 2SPLIT2A to an alias of 2SPLIT1B, see AKNS_fscatter.c. For KdV the issue is now solved by giving 2SPLIT2A the appropriate similarity transformation.