Fixed mistake in residue calculation
In the calculation of the derivative scattering matrix for the CF*_* discretizations, average weights (named scl_factor) were applied instead of their individual "l_weights". This appeared to be the cause for many of the cases where the error in the residue calculation decayed at less than the theoretical rate as a function of the number of samples.