Restored/fixed bounding box calculation nsev.c
For most discretizations the preprocessed `q` contains now weighted averages of subsamples of the original q. The amplification by the upsampling factor is thus removed. For these discretizations there is no need to subsample the preprocessed `q` to estimate the L2 norm of the original q. However, the routine misc_l2norm2() calculates the L2 norm after truncating the staircase reconstruction (midpoint zero order hold) of q by half a sample time at the beginning and half a sample time at the end before calculating the L2 norm. In case `upsampling_factor`>1 one would need to shift the boundaries `a` and `b` a distance 0.5*`eps_t`*(`upsampling_factor`-1)/`upsampling_factor` outwards to be the midpoint of the first and last subsample. A simpler solution is to take the L2 norm of the whole staircase approximation. I left misc_l2norm2() unchanged, because it's also used elsewhere.

Fixed: In the previous implementation the bounding box for the discretizations *(T)ES4 was based on an estimate of || 3 q_t ||_2^2, where q_t is the first time derivative of q and || ||_2 is the L2-norm. These are now based on an estimate of || q ||_2^2 like the other discretizations.