Fixed another mistake in residue calculation
According to Eqs, (5) and (48) in the source of the TES4 discretization, https://doi.org/10.1364/OE.377140 , the one-step change of state matrix is calculated as a product of three matrices, of which the middle one is just the ordinary midpoint rule matrix exponential. The previous implementation was calculating something else and this caused the error decay of the residues with the TES4 discretization to be lower than expected.

Now all error decays for kdvv and nsev without Richardson extrapolation are as expected. For kdvv Richardson extrapolation increases only the order of rho, a, b, and bound states. The only exception is CF5_3, where the order of the norming constants and residues also increase to 6. For nsev with Richardson extrapolation this situation is exactly opposite: All error orders increase by one, except for CF5_3 where the order of the residues and norming constants remains at 5!