Completion new testcases kdvv
Completed the negative rect testcase with true values for a and b.
Added negative rect testcase to all discretizations that don't assume smoothness of the potential, except for 2SPLIT2_modal, for which this testcase runs into a division by 0.
Checked the error bounds with and without FFTW, with GCC-10 and clang compiler.
Renamed the tests to reflect that they test a discretization, with any or more testcases.